[ANI-47-feedback-page] feat: develop branch에 맞춰 페이지 경로 수정
